+++ ft_memcpy.c
@@ -6,15 +6,15 @@
 /*   By: victda-s <victda-s@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 22:34:50 by victda-s          #+#    #+#             */
-/*   Updated: 2024/10/16 12:22:09 by victda-s         ###   ########.fr       */
+/*   Updated: 2024/10/17 15:02:59 by victda-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dest, void *src, int n)
+void	*ft_memcpy(void *dest, void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
